refactor(desktop): extract AssistantMessage pure helpers (GUI monolith run)
11 pure helpers + 2 interfaces + 3 constants (model-ref parsing, provider-route
formatting, tool-state summary, block text extraction, reviewer config, follow-up
gating) move from AssistantMessage.tsx into AssistantMessage.helpers.ts — no React/
hooks/JSX, pure over shared chat types. Component imports the 12 it uses.
10 unit tests (jsdom for the DOMParser path). typecheck + build:renderer clean.

AssistantMessage.tsx 816 -> 671 (under 700).
